+++ src/vbox.h
@@ -1,10 +1,10 @@
 #ifndef VBOX_3D_H
 #define VBOX_3D_H
-
+#include<stddef.h>
 struct Box3d {
-	float sort_param; /* parameter on which to sort the vbox */
-	int n_pix; /* number of pixels in the vbox        */
-	int vol; /* quantized volume of vbox            */
+	size_t sort_param; /* parameter on which to sort the vbox */
+	size_t n_pix; /* number of pixels in the vbox        */
+	size_t vol; /* quantized volume of vbox            */
 	int r1; /* min r index in the vbox             */
 	int r2; /* max r index in the vbox             */
 	int g1; /* min g index in the vbox             */
@@ -22,9 +22,9 @@ extern BOX3D* box_3d_create(int r1, int r2, int g1,
 
 extern BOX3D* box_3d_copy(BOX3D* vbox);
 
-extern int vbox_get_count(BOX3D* vbox, int* histo, int sigbits);
+extern size_t vbox_get_count(BOX3D* vbox, int* histo, int sigbits);
 
-extern int vbox_get_volume(BOX3D* vbox);
+extern size_t vbox_get_volume(BOX3D* vbox);
 extern void vbox_get_average_color(BOX3D* vbox, int* histo,
                                    int sigbits, int index,
                                    int* p_rval, int* p_gval, int* p_bval);
